Region Clicking Implementation
--Added custom QLabel subclass called OverlayLabel, which publically
inherits from QLabel and adds a MousePressedEvent, and emits a signal
called MousePressed(), which allows us to get coordinates clicked inside
of the label
--Updated OverlayView to use the new label, and added some basica
functionality to test the coordinates
--Added two ints: x and y to overlay view, but the x-coord is mainly
just for testing purposes. Feel free to delete later on in development
to clean up our code. This is because we will only be sending the
y-coord.